+++ DatabaseApi/Models/Price.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace DatabaseApi.Models
 {
     public class Price
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public decimal RegularPrice { get; set; }
         public decimal? SalePrice { get; set; }
         public int PriceLevelId { get; set; }
